Create sleep.bat (Hibernate  Command)
Hibernate mode is very similar to sleep, but instead of saving your open documents and running applications to your RAM, it saves them to your hard disk. This allows your computer to turn off entirely, which means once your computer is in Hibernate mode, it uses zero power. Once the computer is powered back on, it will resume everything where you left off. It just takes a bit longer to resume than sleep mode does (though with an SSD, the difference isn’t as noticeable as it is with traditional hard drives).